pyproject.toml
[project]
name = "bulqit-keegan-agent"
version = "0.1.0"
description = "Bulqit Keegan voice agent for homeowner follow-up calls"
requires-python = ">=3.9"
dependencies = [
    "cartesia-line>=0.1.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "config", "llm_node"]
